refactor: result, ranking page 정리
