--- src/containers/ranking/[problemId]/index.ts
@@ -1,1 +1,0 @@
-export { RankingContainer } from './Ranking';
